# public/schema/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rolnopol",
    "version": "1.0.45",
    "description": "API documentation for the Rolnopol service with versioning support",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/your-repo/api-rolnopol"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 (Stable) - Production endpoints"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "token",
        "description": "Authentication token for user endpoints"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Response timestamp"
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Response timestamp"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          }
        }
      },
...
      "UserRegistration": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email for login"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "minLength": 3,
            "maxLength": 50
          },
          "displayedName": {
            "type": "string",
            "description": "Display name (optional)",
            "minLength": 3,
            "maxLength": 20
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email for login"
          },
          "password": {
            "type": "string",
            "description": "User password"
          }
        }
      },
...
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string",
                "description": "Authentication token"
              },
              "id": {
                "type": "integer",
                "description": "Numeric user ID (primary key)"
              },
              "displayedName": {
                "type": "string",
                "description": "User display name"
              },
              "userId": {
                "type": "string",
                "description": "Auto-generated user ID (internal identifier)"
              },
              "email": {
                "type": "string",
                "format": "email",
                "description": "Email for login"
              },
              "expiration": {
                "type": "object",
                "description": "Token expiration information"
              }
            }
          }
        }
      },
...
      "SystemStatistics": {
        "type": "object",
        "properties": {
          "users": {
            "type": "integer",
            "description": "Number of active users"
          },
          "farms": {
            "type": "integer",
            "description": "Total number of farms"
          },
          "area": {
            "type": "number",
            "description": "Total area of all fields"
          },
          "staff": {
            "type": "integer",
            "description": "Total number of staff members"
          },
          "animals": {
            "type": "integer",
            "description": "Total number of animals"
          },
          "avgStaffAge": {
            "type": "integer",
            "description": "Average age of staff members"
          },
          "offers": {
            "type": "integer",
            "description": "Number of active marketplace offers"
          }
        }
      },
...
      "TokenValidation": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Authentication token to validate"
          }
        },
        "required": ["token"]
      },
...
    }
  },
  "paths": {
...
    "/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new user",
        "description": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistration"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate user and get access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/authorization": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user information",
        "description": "Validate token and return current user data",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User information retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Authentication"],
        "summary": "Validate and refresh token",
        "description": "Validate provided token and return user data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TokenValidation" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token validated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/contact": {
      "post": {
        "summary": "Submit contact form",
        "description": "Submit a contact form with name, email, subject, and message.",
        "tags": ["Contact"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "subject": { "type": "string" },
                  "message": { "type": "string" }
                },
                "required": ["name", "email", "subject", "message"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "id": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/statistics": {
      "get": {
        "tags": ["System"],
        "summary": "Get system statistics",
        "description": "Returns comprehensive system statistics including users, farms, fields, staff, animals, and marketplace data",
        "responses": {
          "200": {
            "description": "System statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SystemStatistics" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/map": {
      "get": {
        "tags": ["Map"],
        "summary": "Map API info",
        "description": "Basic info about map endpoints",
        "security": [{ "TokenAuth": [] }],
        "responses": {
          "200": {
            "description": "Info",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } }
          }
        }
      }
    },
    "/map/fieldsmap": {
      "get": {
        "tags": ["Map"],
        "summary": "Get fields geojson",
        "description": "Return fields map JSON (GeoJSON-like)",
        "security": [{ "TokenAuth": [] }],
        "responses": { "200": { "description": "Fields map", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/map/districts": {
      "get": {
        "tags": ["Map"],
        "summary": "List districts (map)",
        "description": "Return district names available in map data",
        "security": [{ "TokenAuth": [] }],
        "responses": {
          "200": {
            "description": "District names",
            "content": { "application/json": { "schema": { "type": "array", "items": { "type": "string" } } } }
          }
        }
      }
    },
...
  }
}